+++ docs/experiments/raven-target/verify_editor.py
@@ -269,7 +269,7 @@ try:
             (29, 'method.', ('Name', 'DeclaringType', 'ReturnType', 'GetParameters', 'IsPublic')),
             (30, 'property.', ('Name', 'CanRead', 'GetGetMethod', 'GetIndexParameters')),
             (31, 'System.Reflection.BindingFlags.', ('Public', 'NonPublic', 'Instance', 'Static', 'DeclaredOnly')),
-            (32, 'System.Array<int>.', ('Allocate', 'View'))):
+            (32, 'System.Runtime.InteropServices.NativeMemory.', ('Alloc', 'Free'))):
             text = 'import System.*\nfunc Main() {\n    let info = typeof(int)\n    let method = info.GetMethods()[0]\n    let property = typeof(Date).GetProperties()[0]\n    ' + expression + '\n}'
             send('textDocument/didChange', {'textDocument': {'uri': uri, 'version': version}, 'contentChanges': [{'text': text}]})
             result = receive(send('textDocument/completion', {'textDocument': {'uri': uri},